+++ diesel/src/associations/mod.rs
@@ -1,8 +1,13 @@
 use std::hash::Hash;
 
+use query_dsl::FindDsl;
+use query_source::Table;
+
 pub trait Identifiable {
     type Id: Hash + Eq + Copy;
+    type Table: Table + FindDsl<Self::Id>;
 
+    fn table() -> Self::Table;
     fn id(&self) -> Self::Id;
 }